upstream: move trace-row helpers + workspace add jolt-host
Make extract_trace_rows / cycle_input / r1cs_cycle_witness pub in
jolt-kernels::trace (along with private fill_next_fields and
lookup_operands helpers). Previously private to
jolt-equivalence::core_oracle, but consumed by any host-side prove
driver (jolt-host, future SDK).

Trade-off: adds common as a jolt-kernels dep (needed for MemoryLayout
in cycle_input). The kernels crate already pulls in jolt-r1cs,
jolt-program, jolt-witness, jolt-riscv, jolt-field — everything else
the helpers need.

core_oracle.rs's local copies are deleted; it imports from kernels now.
Adds workspace member crates/jolt-host with a stub lib.rs (fleshed out
in the next commit).
