+++ crates/jolt-host/src/lib.rs
@@ -1,0 +1,1 @@
+//! Host-side modular prove entry point (WIP — see incoming commits).
